draw: Remove the backing pixmap from score rendering.
Denemo renders the score to a backing pixmap and then
copies it to the widget.  This duplicates work done by
Gtk since it also does back buffering, so there's no
reason to keep it around anymore.  This patch removes
it and adds some rudimentary clipping to the expose
event's damage region.